+++ frontend/src/pages/Profile/ProfileHeader/ProfileHeader.tsx
@@ -31,7 +31,7 @@ return (
 		<img src={profile.avatar} className={styles.profileBackground} />
 		<Grid className={styles.profileMainInfo}>
 			<img src={"http://localhost:8000/" + profile.avatar} className={styles.profileAvatar} />
-			{isOwner && <Button variant="contained" component="label" size="small" color="primary" sx={{ ml: -5 }}>Change avatar
+			{isOwner && <Button variant="contained" component="label" color="primary" sx={{ ml: -5, mr: 2 }}>Change avatar
 				<input type="file" hidden onChange={onAvatarChange} />
 			</Button>}
 			<Grid>